pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nba_application"
version = "0.1.0"
description = "NBA stats analysis application"
authors = [{ name = "Justin DeKock", email = "jdekock17@gmail.com" }]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "astroid==3.3.8",
    "cairocffi==1.7.1",
    "CairoSVG==2.7.1",
    "certifi==2025.1.31",
    "cffi==1.17.1",
    "charset-normalizer==3.4.1",
    "colorama==0.4.6",
    "cssselect2==0.7.0",
    "cx_Oracle==8.3.0",
    "defusedxml==0.7.1",
    "dill==0.3.9",
    "graphviz==0.20.3",
    "idna==3.10",
    "isort==6.0.0",
    "mccabe==0.7.0",
    "numpy==2.2.2",
    "pandas==2.2.3",
    "pillow==11.1.0",
    "platformdirs==4.3.6",
    "pycairo==1.27.0",
    "pycparser==2.22",
    "pylint==3.3.4",
    "python-dateutil==2.9.0.post0",
    "pytz==2025.1",
    "requests==2.32.3",
    "six==1.17.0",
    "tinycss2==1.4.0",
    "tomlkit==0.13.2",
    "tzdata==2025.1",
    "Unidecode==1.3.8",
    "urllib3==2.3.0",
    "webencodings==0.5.1"
]

[tool.setuptools.packages]
find = {}

[project.scripts]
nba_app = "src.main:main"
